--- server/controllers/auth/index.js
@@ -1,9 +1,0 @@
-const { signin } = require('./signin');
-const { signout } = require('./signout');
-const { signup } = require('./signup');
-
-module.exports = {
-  signin,
-  signout,
-  signup,
-};
