--- server/controllers/auth/signout.js
@@ -1,7 +1,0 @@
-// req에서 토큰유무 확인
-// 없으면 로그아웃 요청 거절
-// 있으면 쿠키의 액세스토큰 삭제
-
-// 클라이언트에게 상태로 관리 중인 토큰을 삭제하기 위해 해당 유저의 email 알려주기
-
-// destroyToken
