--- server/controllers/essays/createEssay.js
@@ -1,28 +1,0 @@
-// ! 글 하나에 최대 태그 3개까지 들어감
-
-// 새 글 작성
-// req
-// {
-//   "content": "string",
-//   "isPublic": false,
-//   "tags": [
-//     "string"
-//   ]
-// }
-
-// res
-// {
-//   "status": 201,
-//   "message": "Susscessfully Created New Essay!",
-//   "data": {
-//     "essayId": 0,
-//     "content": "string",
-//     "tags": [
-//       "string"
-//     ],
-//     "isDeleted": false,
-//     "isPublic": false,
-//     "createdAt": "2022-04-01",
-//     "updatedAt": "2022-04-01"
-//   }
-// }
