--- server/controllers/essays/deleteEssay.js
@@ -1,13 +1,0 @@
-// delete(`/trashes/:essayId`, essayCtrl.deleteEssay); // 휴지통의 글 영구 삭제
-//-> essays에서 essayId로 글 삭제
-
-// is_deleted: true -> 휴지통으로 보내기
-
-// req.params.essayId
-// req.cookies.Bearer
-
-// {
-//   "message": "Successfully delete your essay!",
-//   "status": 200,
-//   "permanetlyDeletedEssayId": 0
-// }
