--- server/controllers/essays/index.js
@@ -1,9 +1,0 @@
-const { createEssay } = require('./createEssay');
-const { deleteEssay } = require('./deleteEssay');
-const { updateEssay } = require('./updateEssay');
-
-module.exports = {
-  createEssay,
-  deleteEssay,
-  updateEssay,
-};
