--- server/controllers/space/getEssayListByFilter.js
@@ -1,18 +1,0 @@
-// req.params.filter
-// req.cookies.Bearer
-
-let sampleResponse = {
-  status: 200,
-  data: [
-    {
-      essayId: 0,
-      content: 'string',
-      tags: ['string'],
-      isDeleted: false,
-      isPublic: false,
-      createdAt: '2022-04-01',
-      updatedAt: '2022-04-01',
-    },
-  ],
-  message: 'Successfully read filterd essays with keyword!',
-};
