--- server/controllers/space/getTagList.js
@@ -1,4 +1,0 @@
-// space 화면 좌측에서 사용자의 태그 목록만 반환
-
-// req.cookies의 토큰으로 uuid 구하기
-// email로 userId 구하고
