--- server/controllers/space/getTrashList.js
@@ -1,8 +1,0 @@
-// makeTitleForTrashView
-
-// req.cookies에서 accessToken 해독 후 사용자 메일 얻기
-
-// 사용자의 모든 글 중에서 is_delete:true인 글 목록 조회
-
-// essayId: 글 아이디
-// title: yyyy-mm-dd | 맨 첫 줄 10글자 자르기
