--- server/controllers/space/index.js
@@ -1,11 +1,0 @@
-const { getMySpace } = require('./getMySpace');
-const { getTagList } = require('./getTagList');
-const { getTrshList } = require('./getTrshList');
-const { getEssayListByFilter } = require('./getEssayListByFilter');
-
-module.exports = {
-  getMySpace,
-  getTagList,
-  getTrshList,
-  getEssayListByFilter,
-};
