--- server/controllers/users/deleteUserAccount.js
@@ -1,10 +1,0 @@
-// req.cookies.Bearer
-
-let sampleResponse = {
-  status: 200,
-  message: 'Successfully Delete Your Account!',
-  deletedAccountUserId: 0,
-};
-
-// destoy accessToken
-// delete account from database
