--- server/controllers/users/index.js
@@ -1,9 +1,0 @@
-const { deleteUserAccount } = require('./deleteUserAccount');
-const { getUserInfo } = require('./getUserInfo');
-const { updateUserInfo } = require('./updateUserInfo');
-
-module.exports = {
-  deleteUserAccount,
-  getUserInfo,
-  updateUserInfo,
-};
