--- server/middlewares/auth/createRefreshToken.js
@@ -1,17 +1,0 @@
-const jwt = require('jsonwebtoken');
-const secretKey = process.env.TOKEN_SECRET;
-const expiresIn = process.env.REF_EXPIRED_IN;
-
-// user는 객체로 회원의 정보를 담는 객체
-// ! user의 유효성 검사가 필요함
-const createRefreshToken = (accountInfoForToken = {}) => {
-  const { email, username, profileImage /*uuid,*/ } = accountInfoForToken;
-  const payload = {
-    email,
-    username,
-    profileImage,
-  };
-  return jwt.sign(payload, secretKey, { expiresIn });
-};
-
-module.exports = { createRefreshToken };
