--- server/middlewares/auth/destroyToken.js
@@ -1,1 +1,0 @@
-// 요청 시 응답 쿠키의 토큰 찾고 파괴 → 클라이언트한테 uuid 반환하고 해당 유저의 상태변수로 관리 중인 accessToken 지우도록
