--- server/middlewares/formatter/dateFormatter.js
@@ -1,2 +1,0 @@
-// user.birthday, essay.created_at, essay.updated_at, record.markList
-// yyyy-mm-dd
