--- server/middlewares/formatter/index.js
@@ -1,7 +1,0 @@
-const { dateFormatter } = require('./dateFormatter');
-const { makeTitleForTrashList } = require('./makeTitleForTrashList');
-
-module.exports = {
-  dateFormatter,
-  makeTitleForTrashList,
-};
