--- server/middlewares/utils/index.js
@@ -1,7 +1,0 @@
-const { encrypt } = require('./encrypt');
-const { decrypt } = require('./decrypt');
-
-module.export = {
-  encrypt,
-  decrypt,
-};
