--- server/routes/landing.js
@@ -1,6 +1,0 @@
-const router = require('express').Router();
-const landingCtrl = require('../controllers/landing');
-
-router.get('/', landingCtrl.getLanding);
-
-module.exports = router;
