--- server/routes/tag.js
@@ -1,6 +1,0 @@
-const router = require('express').Router();
-const { getTagList } = require('../controllers/space');
-
-router.get('/', getTagList);
-
-module.exports = router;
